requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.31
pydantic==1.10.2
python-dotenv==1.0.1